Use O2 optimization and no -march flag
RPM builds on linux generally use a default optimization level of
O2, so use this in place of Homebrew's Os. These only differ for
gcc in terms of Os removing any optimizations that would increase
the size of the binary.

Do not add any -march flags. We may have several different CPUs
on the desktop, even if all are Intel so don't specialize.