Build CLHEP with explicit install RPATH
Though homebrew will supply the correct final rpath through LDFLAGS,
CMake builds appear to ignore this for certain libraries and for
not understood reasons.

As a temporary workaround, add CMAKE_BUILD_WITH_INSTALL_RPATH to
args when building CLHEP. This prevents stripping of the rpath in
te final install, though may make testing as part of the install
step problematic.